move some stuff into `ExternalLinkEmbed` instead
